Amelioration conversion type/dynamic_cast dans fonction. + entier.val devient long int

+++ ProjetLO21/entier.h
@@ -6,7 +6,7 @@
 
 class Entier : public Nombre {
 private :
-    double val;
+    long int val;
 
 public :
     Entier(double v=0) : val(v) {}
